server/requirements-base.txt
asgiref==3.8.1
celery==5.5.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
psycopg2==2.9.10
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
